16 move hooks logic to dedicated file
