+++ .gitignore
@@ -2,5 +2,3 @@
 oas/*
 response/*
 .vscode/*
-hooks*
-hooks/*
