+++ hooks.yml
@@ -1,0 +1,4 @@
+- path: /comm/v2/flight-offer/flight-search
+  method: POST
+  body:
+    flightSearch.departureLocationCode.airportCode: AMS
